switching some values from size_t to uint64_t so that they're a guarenteed size when saving/loading

+++ hexmap/src/data/hex_map.h
@@ -31,12 +31,12 @@ namespace data
 
     struct hxm_header_t
     {
-        size_t version{0}; //this might need to be seperate from the header in case the header changes
+        uint64_t version{0}; //this might need to be seperate from the header in case the header changes
 
         glm::uvec2 map_size{0};
         glm::uvec2 chunk_size{0};
-        size_t num_map_objects{0};
-        size_t num_splines{0};
+        uint64_t num_map_objects{0};
+        uint64_t num_splines{0};
     };
 }
 }
